test-requirements.txt
coverage
mock
pep8
nose
python-ldap
